Rename client file
